fix some paths from the parts-components rename
