+++ views/menus/menus/bonapp-menu.js
@@ -3,14 +3,14 @@ import React from 'react'
 import {View, Text} from 'react-native'
 import LoadingView from '../../components/loading'
 import type {TopLevelViewPropsType} from '../../types'
-import {FilteredMenuView} from '../parts/filtered-menu'
+import {FilteredMenuView} from '../components/filtered-menu'
 import type {BonAppMenuInfoType, BonAppCafeInfoType} from '../types'
 import {bonappMenuBaseUrl, bonappCafeBaseUrl} from '../data'
 import sample from 'lodash/sample'
 import type momentT from 'moment'
 import moment from 'moment-timezone'
 const CENTRAL_TZ = 'America/Winnipeg'
-import {findMenu} from './find-menu'
+import {findMenu} from '../lib/find-menu'
 import {fetchJson} from '../../components/fetch'
 
 export class BonAppHostedMenu extends React.Component {
